+++ okitweb/static/okit/js/oci_artefacts/compartment.js
@@ -118,13 +118,7 @@ class Compartment extends OkitContainerArtifact {
      */
     draw() {
         console.groupCollapsed('Drawing ' + Compartment.getArtifactReference() + ' : ' + this.id);
-        let svg = drawArtifact(this.getSvgDefinition());
-        // Add Properties Load Event to created svg
-        let me = this;
-        svg.on("click", function() {
-            me.loadProperties();
-            d3.event.stopPropagation();
-        });
+        let svg = super.draw();
         console.groupEnd();
     }