+++ okitweb/static/okit/js/oci_artefacts/route_table.js
@@ -61,17 +61,8 @@ class RouteTable extends OkitArtifact {
             console.groupEnd();
             return;
         }
-        let svg = drawArtifact(this.getSvgDefinition());
-        /*
-        ** Add Properties Load Event to created svg. We require the definition of the local variable "me" so that it can
-        ** be used in the function dur to the fact that using "this" in the function will refer to the function not the
-        ** Artifact.
-         */
         let me = this;
-        svg.on("click", function() {
-            me.loadProperties();
-            d3.event.stopPropagation();
-        });
+        let svg = super.draw();
         console.groupEnd();
     }